--- customers.csv
@@ -1,1 +1,0 @@
-1001,John,Smith
