--- src/Account_Classes/Account.java
@@ -1,21 +1,0 @@
-package Account_Classes;
-import User_Classes.*;
-
-public abstract class Account{
-    String accountNumber;
-    private double balance;
-    customer Owner;
-
-    public void deposit(double amount){
-        System.out.println("Successfully Deposited "+ amount);
-        System.out.print(this.balance);
-        this.balance += amount;
-        System.out.print(" → "+ this.balance+"\n");
-    }
-    public void withdraw(double amount){
-        this.balance -= amount; // CS: just put this in temporarily to test transfer cause I didn't realize this didn't do anyhting
-    }
-    public double getBalance(){
-        return this.balance;
-    }
-}
